__app__/requirements.txt
azure-functions==1.3.0
backoff==1.10.0
loguru==0.5.1
PyJWT==1.7.1
pymongo==3.11.0
